+++ NEWS.md
@@ -1,4 +1,9 @@
-# User visible changes for EasyFITS
+# User visible changes for `EasyFITS`
+
+This page describes the most important changes in `EasyFITS`. The format is based on [Keep a
+Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic
+Versioning](https://semver.org).
+
 
 ## Unreleased