+++ NEWS.md
@@ -4,8 +4,7 @@ This page describes the most important changes in `EasyFITS`. The format is base
 Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic
 Versioning](https://semver.org).
 
-
-## Unreleased
+## Version 0.7.2 [2026-01-08]
 
 ### Fixed